+++ src/DepthCameraSensor.cc
@@ -31,7 +31,8 @@
 #include "ignition/sensors/DepthCameraSensor.hh"
 #include "ignition/sensors/Manager.hh"
 #include "ignition/sensors/SensorFactory.hh"
-#include "ignition/sensors/GaussianNoiseModel.hh"
+#include "ignition/sensors/ImageGaussianNoiseModel.hh"
+#include "ignition/sensors/ImageNoise.hh"
 #include "ignition/sensors/RenderingEvents.hh"
 
 #include "PointCloudUtil.hh"
@@ -347,7 +348,7 @@ bool DepthCameraSensor::CreateCamera()
     if (noiseSdf.Type() == sdf::NoiseType::GAUSSIAN)
     {
       this->dataPtr->noises[noiseType] =
-        NoiseFactory::NewNoiseModel(noiseSdf, "depth");
+        ImageNoiseFactory::NewNoiseModel(noiseSdf, "depth");
 
       std::dynamic_pointer_cast<ImageGaussianNoiseModel>(
            this->dataPtr->noises[noiseType])->SetCamera(