+++ src/GaussianNoiseModel.cc
@@ -25,11 +25,6 @@
 #include <ignition/math/Rand.hh>
 
 #include "ignition/common/Console.hh"
-#include "ignition/rendering/GaussianNoisePass.hh"
-#include "ignition/rendering/RenderPass.hh"
-#include "ignition/rendering/RenderEngine.hh"
-#include "ignition/rendering/RenderPassSystem.hh"
-
 
 using namespace ignition;
 using namespace sensors;
@@ -61,20 +56,6 @@ class ignition::sensors::GaussianNoiseModelPrivate
 
   /// \brief True if the type is GAUSSIAN_QUANTIZED
   public: bool quantized = false;
-};
-
-class ignition::sensors::ImageGaussianNoiseModelPrivate
-{
-  /// \brief If type starts with GAUSSIAN, the mean of the distribution
-  /// from which we sample when adding noise.
-  public: double mean = 0.0;
-
-  /// \brief If type starts with GAUSSIAN, the standard deviation of the
-  /// distribution from which we sample when adding noise.
-  public: double stdDev = 0.0;
-
-  /// \brief Gaussian noise pass.
-  public: rendering::GaussianNoisePassPtr gaussianNoisePass;
 };
 
 //////////////////////////////////////////////////
@@ -194,58 +175,4 @@ void GaussianNoiseModel::Print(std::ostream &_out) const
     << "bias[" << this->dataPtr->bias << "] "
     << "precision[" << this->dataPtr->precision << "] "
     << "quantized[" << this->dataPtr->quantized << "]";
-}
-
-//////////////////////////////////////////////////
-ImageGaussianNoiseModel::ImageGaussianNoiseModel()
-  : GaussianNoiseModel(), dataPtr(new ImageGaussianNoiseModelPrivate())
-{
-}
-
-//////////////////////////////////////////////////
-ImageGaussianNoiseModel::~ImageGaussianNoiseModel()
-{
-  delete this->dataPtr;
-  this->dataPtr = nullptr;
-}
-
-//////////////////////////////////////////////////
-void ImageGaussianNoiseModel::Load(const sdf::Noise &_sdf)
-{
-  Noise::Load(_sdf);
-
-  this->dataPtr->mean = _sdf.Mean();
-  this->dataPtr->stdDev = _sdf.StdDev();
-}
-
-//////////////////////////////////////////////////
-void ImageGaussianNoiseModel::SetCamera(rendering::CameraPtr _camera)
-{
-  if (!_camera)
-  {
-    ignerr << "Unable to apply gaussian noise, camera is null\n";
-    return;
-  }
-
-  rendering::RenderEngine *engine = _camera->Scene()->Engine();
-  rendering::RenderPassSystemPtr rpSystem = engine->RenderPassSystem();
-  if (rpSystem)
-  {
-    // add gaussian noise pass
-    rendering::RenderPassPtr noisePass =
-      rpSystem->Create<rendering::GaussianNoisePass>();
-    this->dataPtr->gaussianNoisePass =
-        std::dynamic_pointer_cast<rendering::GaussianNoisePass>(noisePass);
-    this->dataPtr->gaussianNoisePass->SetMean(this->dataPtr->mean);
-    this->dataPtr->gaussianNoisePass->SetStdDev(this->dataPtr->stdDev);
-    this->dataPtr->gaussianNoisePass->SetEnabled(true);
-    _camera->AddRenderPass(this->dataPtr->gaussianNoisePass);
-  }
-}
-
-//////////////////////////////////////////////////
-void ImageGaussianNoiseModel::Print(std::ostream &_out) const
-{
-  _out << "Image Gaussian noise, mean[" << this->dataPtr->mean << "], "
-    << "stdDev[" << this->dataPtr->stdDev << "] ";
 }